# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zytrixon Mobile API Documentation",
        "description": "API documentation for the Zytrixon Truck Billing mobile application.",
        "contact": {
            "email": "admin@zytrixon.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Get all trips",
                "description": "Retrieve a paginated list of trips based on the authenticated user's role.",
                "operationId": "edefb369ed6527e9afd948b101a74b9f",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by location, truck, or ID",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Start date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "End date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trips retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "Create a new trip",
                "operationId": "c27804b77dfcd9567f2dba32b2931dad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_id",
                                    "driver_id",
                                    "from_location",
                                    "to_location",
                                    "start_date"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "driver_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "dealer_id": {
                                        "type": "integer",
                                        "example": 3,
                                        "nullable": true
                                    },
                                    "from_location": {
                                        "type": "string",
                                        "example": "DELHI"
                                    },
                                    "to_location": {
                                        "type": "string",
                                        "example": "MUMBAI"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Trip created successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trips/{id}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Get specific trip details",
                "operationId": "73566c003dc292ba95537ba7c85c3889",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip details retrieved successfully"
                    },
                    "404": {
                        "description": "Trip not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trips/{id}/status": {
            "put": {
                "tags": [
                    "Trips"
                ],
                "summary": "Update trip status",
                "operationId": "9b0c3f7ad72b053af7b6d74520cb20ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "scheduled",
                                            "in_transit",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trips/{id}/end": {
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "End a trip",
                "description": "Marks trip as completed and releases the vehicle.",
                "operationId": "59d23b63fda895b308d8e28c9b90824f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip ended successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token here",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}